+++ changelog.md
@@ -4,6 +4,7 @@
 * Resolved an issue with MKV Legacy output (wrong cast exception). 
 * Zoom and pan effects perfomance optimizations
 * Added basic Media Blocks API (WIP)
+* Added HLS network streaming to Video Edit SDK .Net
 
 15.2
 * Added HorizontalAlignment and VerticalAlignment properties to the text and image logos